broker/pyproject.toml
[tool.poetry]
name = "broker"
version = "0.1.0"
description = ""
authors = ["Lorenzo Carbonell <lorenzo.carbonell.cerezo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
lxml = "^4.9.3"
cssselect = "^1.2.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
